refactor: make cargo api more consistant
also moves component stuff to a separate file
