+++ internal/commands/find_release_version.go
@@ -43,9 +43,9 @@ func (cmd *FindReleaseVersion) Execute(args []string) error {
 	}
 	releaseSource := cmd.mrsProvider(kilnfile, false)
 
-	spec, ok := kilnfile.ComponentSpec(cmd.Options.Release)
-	if !ok {
-		return cargo.ErrorSpecNotFound(cmd.Options.Release)
+	spec, err := kilnfile.FindReleaseWithName(cmd.Options.Release)
+	if err != nil {
+		return err
 	}
 
 	spec.StemcellOS = kilnfileLock.Stemcell.OS
